Testing numbers and f32

+++ fuzz/fuzz_targets/parser_float_32.rs
@@ -11,7 +11,7 @@ fn check_roundtrip_default(input: Value, printed: &str) {
 }
 
 
-fuzz_target!(|input: f64| {
+fuzz_target!(|input: f32| {
     
     let input_as_string: &str = &*input.to_string();
